requirements.txt
azure-functions
azure-storage-blob
langchain
PyPDF2
python-docx
langchain-community
langchain-text-splitters
docx2txt
openai
azure-data-tables
pyjwt
bcrypt
pytest
langchain_experimental 
langchain_postgres
langchain_openai
psycopg